+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-26T11:00:17.523Z",
+    "2026-02-26T12:31:53.721Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {