+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-26T11:00:13.826Z",
+    "2026-02-26T12:31:50.204Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {