+++ packages/orq-rc/RELEASES.md
@@ -7329,3 +7329,13 @@ Based on:
 - [typescript v4.5.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.12 - packages/orq-rc
+
+## 2026-02-26 12:34:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.730.1 (2.844.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.13 - packages/orq-rc